slight refactor.
to find the bug, let's run an even larger test case directly through test code.
this is important is because, when running the code in main.rs against
an actual file, i am seeing that duplicate keys are identified while
inverting the prefix table. This means the tree isn't being encoded or
decoded correctly