prefixes might be longer than 8 bits
I think this is the cause of the bug. I thought capital letters were
related, and they are, but indirectly. Low frequency character are more
likely to have their leaves in the Huffman tree at a lower depth. This
results in longer prefix codes. They are more likely than higher freq
ones to be garbled, which is what i've observed.